Moved HoloViews pane tests into separate module
